+++ app/sprinkles/core/locale/fa/errors.php
@@ -1,0 +1,52 @@
+<?php
+
+/**
+ * UserFrosting (http://www.userfrosting.com)
+ *
+ * @link      https://github.com/userfrosting/UserFrosting
+ * @license   https://github.com/userfrosting/UserFrosting/blob/master/licenses/UserFrosting.md (MIT License)
+ *
+ * Standard Farsi/Persian message token translations for the 'core' sprinkle.
+ *
+ * @package userfrosting\i18n\fa
+ * @author aminakbari
+ */
+
+return [
+    "ERROR" => [
+        "@TRANSLATION" => "خطا",
+
+        "400" => [
+            "TITLE" => "خطا 400: درخواست بد",
+            "DESCRIPTION" => "احتمالا تقصیر شما نیست.",
+        ],
+
+        "404" => [
+            "TITLE" => "خطا 404: صفحه یافت نشد",
+            "DESCRIPTION" => "به نظر نمیرسد چیزی را که دنبالش هستید پیدا کنیم.",
+            "DETAIL" => "ما سعی کردیم صفحه شما را پیدا کنیم...",
+            "EXPLAIN" => "ما نتوانستیم صفحه ی مورد نظر شما را پیدا کنیم.",
+            "RETURN" => 'در هر حال، <a href="{{url}}">اینجا</a> کلیک کنید تا به صفحه اصلی بازگردید.'
+        ],
+
+        "CONFIG" => [
+            "TITLE" => "خطای تنظیمات یوزرفروستینگ!",
+            "DESCRIPTION" => "برخی از الزامات پیکربندی یوزرفروستینگ به نتیجه نرسید.",
+            "DETAIL" => "خطایی پیش آمد.",
+            "RETURN" => 'لطفا خطاهای زیر را اصلاح کنید و سپس مجددا <a href="{{url}}">بارگذاری</a> نمایید.'
+        ],
+
+        "DESCRIPTION" => "ما یک اختلال بزرگ در سیستم احساس کردیم.",
+        "DETAIL" => "این چیزی است که ما دریافت کردیم",
+
+        "ENCOUNTERED" => "خطایی ویژه. نمیدانیم مشکل چیست.",
+
+        "MAIL" => "خطا در ارسال ایمیل. لطفا با مدیر سیستم تماس برقرار کنید. لطفا لاگ را بررسی کنید.",
+        
+        "RETURN" => '<a href="{{url}}">اینجا</a> کلیک تا به صفحه اصلی بازگردید.',
+
+        "SERVER" => "به نظر می آید که در سرور خطایی بوجود آمد. لطفا لاگ پی اچ پی و یوزرفروستینگ را چک کنید.",
+
+        "TITLE" => "اختلالی پدید آمد."
+    ]
+];