+++ app/system/Sprinkle/SprinkleManager.php
@@ -168,7 +168,7 @@ class SprinkleManager
     }
 
     /**
-     * Initialize all base sprinkles in a specified sprinkles.json schema file.
+     * Initialize all base sprinkles in a specified Sprinkles schema file (e.g. 'sprinkles.json').
      *
      * @param string $schemaPath
      */
@@ -217,7 +217,7 @@ class SprinkleManager
     }
 
     /**
-     * Load list of Sprinkles from a JSON schema file (e.g. sprinkles.json).
+     * Load list of Sprinkles from a JSON schema file (e.g. 'sprinkles.json').
      *
      * @param string $schemaPath
      * @return string[]