+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## v4.1.15-alpha
+- Refactor `Password` into a instantiable `Hasher` class, service, and `Password` facade (#827)
+- Change default hash cost back to 10 and fix legacy hash detection issue
+
+## v4.1.14-alpha
+- Fix issue with scopes being applied twice in `Unique::getPaginatedQuery` (https://github.com/userfrosting/extend-user/issues/2)
+- Update Bower dependencies in core Sprinkle
+- Refactor the `Password` class to use `hash_equals` for legacy passwords (prevent timing-based attacks) and factor out the default cost (#814)
+- Check if `require_email_verification` is set in `Authenticator` and sign-in page (#815)
+- Factor out hardcoded `sprinkles.json` filename (partially addresses #813)
+- Add Farsi translations (#816)
+- `ufTable`: Make `tableElement` configurable, and check for existence (#824)
+- Put AdminLTE menu toggle button back (Revert e8a26fb and part of a46205f)
+
 ## v4.1.13-alpha
 - `ufTable`: Implement `rowTemplate` for customizing how rows are rendered (#787)
 - `ufTable`: Support for passing callbacks for column templates instead of Handlebars templates